tests pass.  migrates Operators to use the new SuperMapWrapper, then renamed SuperMapWrapper to SuperMap, since it is the only thing Operators use at this point.  refs #291
