+++ src/data_structures/SuperMap.hh
@@ -7,135 +7,96 @@
   Author: Ethan Coon (ecoon@lanl.gov)
 */
 
-//! SuperMap class for blocking same-type maps into a single, common map.
+//! SuperMap class provides a convenient way of creating and using SuperMapLumped
 
-/*!
-  Takes non-contiguous data structure spaces (CompositeVector, TreeVector)
-  and converts them into a single map.
+/*
+  Amanzi uses SuperMapLumped in a few different ways that make its natural interface
+  not that convenient.  SuperMapLumped also has a few limitations that simplify its
+  design and implementaiton a great deal.
 
-  DESIGN FLAW: Currently this assumes that component names are unique, and if
-  two components share the same name, they share the same map.  This is
-  obviously wrong when multple meshes are involved -- for instance a TV of
-  surface + subsurface, both with "cell" components, would break miserably.
+  This class is a Helper class in that it wraps a SuperMapLumped, providing a
+  related interface that is better designed for users.
 
-  That said, this is a simple and well-posed class.  It stays, but users
-  should prefer to use the wrapper class, SuperMapWrapper, which deals with a
-  few shortcomings of this class.
+  It also enforces and mitigates the design limitations of SuperMapLumped itself.
 */
 
-#ifndef AMANZI_OPERATORS_SUPER_MAP_HH_
-#define AMANZI_OPERATORS_SUPER_MAP_HH_
+#ifndef AMANZI_OPERATORS_SUPER_MAP_WRAPPER_HH_
+#define AMANZI_OPERATORS_SUPER_MAP_WRAPPER_HH_
 
+#include <map>
 #include "Teuchos_RCP.hpp"
-#include "AmanziComm.hh"
+#include "Epetra_Map.h"
+#include "Epetra_BlockMap.h"
 
-#include "dbc.hh"
-#include "Mesh.hh"
-
+#include "SuperMapLumped.hh"
 
 namespace Amanzi {
+
+class CompositeVectorSpace;
+class TreeVectorSpace;
+
+namespace AmanziMesh {
+class Mesh;
+}
+
+
 namespace Operators {
 
+// wrapper class
 class SuperMap {
  public:
-  // Constructor
-  SuperMap(const Comm_ptr_type& comm,
-           const std::vector<std::string>& compnames,
-           const std::vector<int>& dofnums,
-           const std::vector<Teuchos::RCP<const Epetra_BlockMap> >& maps,
-           const std::vector<Teuchos::RCP<const Epetra_BlockMap> >& ghost_maps);
-
-  SuperMap(const SuperMap& other) = delete;
-  virtual ~SuperMap() = default;
-
-  // meta-data
-  bool HasComponent(const std::string& compname) const;
+  SuperMap(const std::vector<CompositeVectorSpace>& cvss);
 
   // map accessors
-  Teuchos::RCP<const Epetra_Map> Map() const { return map_; }
-  Teuchos::RCP<const Epetra_Map> GhostedMap() const { return ghosted_map_; }
+  // -- global map accessors
+  Teuchos::RCP<const Epetra_Map> Map() const { return smap_->Map(); }
+  Teuchos::RCP<const Epetra_Map> GhostedMap() const { return smap_->GhostedMap(); }
 
   // -- component map accessors
   Teuchos::RCP<const Epetra_BlockMap>
-  ComponentMap(const std::string& compname) const {
-    return comp_maps_.at(compname);
+  ComponentMap(int block_num, const std::string& compname) const {
+    return smap_->ComponentMap(block_info_.at(std::make_tuple(block_num, compname, 0)).first);
   }
 
   Teuchos::RCP<const Epetra_BlockMap>
-  ComponentGhostedMap(const std::string& compname) {
-    return comp_ghosted_maps_.at(compname);
+  ComponentGhostedMap(int block_num, const std::string& compname) const {
+    return smap_->ComponentGhostedMap(block_info_.at(std::make_tuple(block_num, compname, 0)).first);
   }
   
   // index accessors
-  const std::vector<int>& Indices(const std::string& compname, int dofnum) const;
-  const std::vector<int>& GhostIndices(const std::string& compname, int dofnum) const;
+  const std::vector<int>&
+  Indices(int block_num, const std::string& compname, int dof_num) const {
+    auto bi = block_info_.at(std::make_tuple(block_num, compname, dof_num));
+    return smap_->Indices(bi.first, bi.second);
+  }
+
+  const std::vector<int>&
+  GhostIndices(int block_num, const std::string& compname, int dof_num) const {
+    auto bi = block_info_.at(std::make_tuple(block_num, compname, dof_num));
+    return smap_->GhostIndices(bi.first, bi.second);
+  }
 
   // block indices.  This is an array of integers, length Map().MyLength(),
   // where each dof and component have a unique integer value.  The returned
   // int is the number of unique values, equal to
   // sum(NumDofs(comp) for comp in components), in this array.
-  std::pair<int, Teuchos::RCP<std::vector<int> > > BlockIndices() const;
-
-#ifdef SUPERMAP_TESTING
- public:
-#else
- protected:
-#endif
+  std::pair<int, Teuchos::RCP<std::vector<int> > > BlockIndices() const {
+    return smap_->BlockIndices();
+  }
   
-  // meta-data accessors
-  int Offset(const std::string& compname) const { return offsets_.at(compname); }
-  int GhostedOffset(const std::string& compname) const { return ghosted_offsets_.at(compname); }
-  int NumOwnedElements(const std::string& compname) const { return counts_.at(compname); }
-  int NumUsedElements(const std::string& compname) const {
-    return counts_.at(compname) + ghosted_counts_.at(compname); }
-  int NumDofs(const std::string& compname) const { return num_dofs_.at(compname); }
-
-  // iterate over compnames
-  typedef std::vector<std::string>::const_iterator name_iterator;
-  name_iterator begin() const { return compnames_.begin(); }
-  name_iterator end() const { return compnames_.end(); }
-  unsigned int size() const { return compnames_.size(); }
-
  protected:
-
-  // Constructs and returns the vector of indices for a given component into the SuperMap
-  virtual const std::vector<int>& CreateIndices_(const std::string& compname, int dofnum, bool ghosted) const;
-
-  // step one of the construction process
-  //
-  // After this, CreateIndices_() can be called
-  void CreateIndexing_();
-
-  // step two of the construction process
-  //
-  // This creates the SuperMap and uses CreateIndices_() to populate and
-  // create the ghosted SuperMap.
-  void CreateMap_(const Comm_ptr_type& comm);
-
- protected:
-  std::vector<std::string> compnames_;
-  std::map<std::string,int> offsets_;
-  std::map<std::string,int> num_dofs_;
-  std::map<std::string,int> counts_;
-  std::map<std::string,int> ghosted_offsets_;
-  std::map<std::string,int> ghosted_counts_;
-
-  int n_local_;
-  int n_local_ghosted_;
-  
-  mutable std::map<std::string, std::map<int, std::vector<int> > > indices_;
-  mutable std::map<std::string, std::map<int, std::vector<int> > > ghosted_indices_;
-
-  Teuchos::RCP<Epetra_Map> map_;
-  Teuchos::RCP<Epetra_Map> ghosted_map_;
-
-  std::map<std::string, Teuchos::RCP<const Epetra_BlockMap> > comp_maps_;
-  std::map<std::string, Teuchos::RCP<const Epetra_BlockMap> > comp_ghosted_maps_;
-
+  std::unique_ptr<SuperMapLumped> smap_;
+  std::map< std::tuple<int,std::string,int>, std::pair<std::string,int> > block_info_;
 };
 
 
-} // namespace Operators
-} // namespace Amanzi
+// Nonmember contructors/factories
+Teuchos::RCP<SuperMap> createSuperMap(const CompositeVectorSpace& cv);
+Teuchos::RCP<SuperMap> createSuperMap(const TreeVectorSpace& cv);
+
+} // namespace
+} // namespace
+
+
 
 #endif
