+++ src/operators/Operator_ConsistentFace.cc
@@ -80,8 +80,8 @@ void Operator_ConsistentFace::SymbolicAssembleMatrixOp(const Op_Cell_FaceCell& o
   std::vector<int> lid_c(cell_max_faces);
 
   // ELEMENT: cell, DOFS: cell and face
-  const std::vector<int>& face_row_inds = map.GhostIndices("face", my_block_row);
-  const std::vector<int>& face_col_inds = map.GhostIndices("face", my_block_col);
+  const std::vector<int>& face_row_inds = map.GhostIndices(my_block_row, "face", 0);
+  const std::vector<int>& face_col_inds = map.GhostIndices(my_block_col, "face", 0);
 
   int ierr(0);
   AmanziMesh::Entity_ID_List faces;
@@ -114,8 +114,8 @@ void Operator_ConsistentFace::AssembleMatrixOp(const Op_Cell_FaceCell& op,
   std::vector<double> vals(cell_max_faces);
 
   // ELEMENT: cell, DOFS: face and cell
-  const std::vector<int>& face_row_inds = map.GhostIndices("face", my_block_row);
-  const std::vector<int>& face_col_inds = map.GhostIndices("face", my_block_col);
+  const std::vector<int>& face_row_inds = map.GhostIndices(my_block_row, "face", 0);
+  const std::vector<int>& face_col_inds = map.GhostIndices(my_block_col, "face", 0);
 
   int ierr(0);
   AmanziMesh::Entity_ID_List faces;
